Fixed get_roles id type

+++ db/init/311_FUNC_get_roles.sql
@@ -1,6 +1,6 @@
 create or replace function get_roles()
 returns table (
-    id integer,
+    id smallint,
     name character varying,
     description text,
     active boolean,
